requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
openai>=1.35.0
python-pptx>=0.6.22
oletools>=0.60.2
httpx>=0.27.0
starlette>=0.37.2
pdf2image>=1.17.0
